+++ ikalog/IkaOutput_Slack.py
@@ -18,7 +18,7 @@
 #  limitations under the License.
 #
 
-from IkaUtils import *
+from .IkaUtils import *
 
 # Needed in GUI mode
 try:
