+++ ikalog/IkaScene_GameStart.py
@@ -18,11 +18,7 @@
 #  limitations under the License.
 #
 
-import numpy as np
-import cv2
-import time
-import sys
-from IkaUtils import *
+from .IkaUtils import *
 
 
 class IkaScene_GameStart:
