+++ ikalog/inputs/__init__.py
@@ -1,0 +1,1 @@
+from ikalog.inputs import cvcapture
