--- input/__init__.py
@@ -1,1 +1,0 @@
-from input.cvcapture import cvcapture
